03_bs calc 메소드 추가

+++ board/src/com/board/common/Test.java
@@ -3,6 +3,6 @@ package com.board.common;
 public interface Test {
 	void test();
 	String info();
-	int calc(double a, double b);
+	double calc(double a, double b);
 	
 }
